+++ R/CreateEvaluationCohort.R
@@ -76,6 +76,7 @@
 #'                                         to start ineligible included visits
 #' @param exclusionEvaluationDaysFromEnd   The number of days from the cohort end date of the exclusionEvaluationCohortId
 #'                                         to end ineligible included visits
+#' @param priorModelToUse                  folder where a previously developed model to use in analysis will be found
 #' @param minimumOffsetFromStart           Minimum number of days to offset for the analysis visit from the start of the observation period
 #' @param minimumOffsetFromEnd             Minimum number of days to offset for the analysis visit from the end of the observation period
 #' @param modelBaseSampleSize              The number of non-xSpec subjects to include in the model
@@ -131,6 +132,7 @@ createEvaluationCohort <- function(connectionDetails,
                                    exclusionEvaluationCohortId = 0,
                                    exclusionEvaluationDaysFromStart = 0,
                                    exclusionEvaluationDaysFromEnd = 0,
+                                   priorModelToUse = NULL,
                                    minimumOffsetFromStart = 365,
                                    minimumOffsetFromEnd = 365,
                                    modelBaseSampleSize = 25000,
@@ -205,6 +207,7 @@ createEvaluationCohort <- function(connectionDetails,
     modelPopulationCohortId = modelPopulationCohortId,
     modelPopulationCohortIdStartDay = modelPopulationCohortIdStartDay,
     modelPopulationCohortIdEndDay = modelPopulationCohortIdEndDay,
+    priorModelToUse = priorModelToUse,
     modelBaseSampleSize = modelBaseSampleSize,
     minimumOffsetFromStart = minimumOffsetFromStart,
     minimumOffsetFromEnd = minimumOffsetFromEnd,